finish cliant refactor
