+++ client/src/App.js
@@ -1,9 +1,7 @@
 import './App.css';
 import { useEffect, useState } from 'react';
-import { BrowserRouter, Route, Routes } from "react-router-dom";
 import UserConnected from './context/UserConnected';
 import Home from './componets/home/Home';
-import About from './componets/about/About'
 import CssBaseline from '@mui/material/CssBaseline';
 import { Experimental_CssVarsProvider as CssVarsProvider } from '@mui/material/styles';
 import axios from 'axios';
@@ -62,12 +60,6 @@ function App() {
         <CssBaseline>
           <UserConnected.Provider value={{ userConnected, setUserConnected }}>
             <Home />
-            {/* <BrowserRouter>
-            <Routes>
-                <Route path='/' element={<Home /> }/>
-                <Route path='/about' element={<About /> }/>
-            </Routes>
-          </BrowserRouter> */}
           </UserConnected.Provider>
         </CssBaseline>
       </CssVarsProvider>
