+++ client/src/componets/authentication/auth/Auth.jsx
@@ -4,7 +4,13 @@ import LogIn from "../logIn/LogIn";
 import ForgetPassword from "../forgetPassword/ForgetPassword";
 import ResetPassword from "../resetPassword/ResetPassword";
 
-export default function Auth({ handleCloseLogIn }) {
+import Dialog from "@mui/material/Dialog";
+import DialogContent from "@mui/material/DialogContent";
+import DialogTitle from "@mui/material/DialogTitle";
+import IconButton from "@mui/material/IconButton";
+import CloseIcon from "@mui/icons-material/Close";
+
+export default function Auth({ handleCloseLogIn, openLogIn }) {
   const [registered, setRegistered] = useState(true);
   const [havePass, setHavePass] = useState(true);
   const [resetPass, setResetPass] = useState(false);
@@ -20,22 +26,44 @@ export default function Auth({ handleCloseLogIn }) {
   };
 
   return (
-    <div className="Auth">
-      {registered ? (
-        havePass ? (
-          <LogIn
-            handleRegistered={handleRegistered}
-            handleHavePass={handleHavePass}
-            handleCloseLogIn={handleCloseLogIn}
-          />
-        ) : resetPass ? (
-          <ResetPassword handleHavePass={handleHavePass} />
-        ) : (
-          <ForgetPassword handleResetPass={handleResetPass} />
-        )
-      ) : (
-        <Register handleRegistered={handleRegistered} />
-      )}
+    <div>
+      <Dialog
+        open={openLogIn}
+        onClose={handleCloseLogIn}
+        aria-describedby="auth"
+      >
+        <DialogTitle>
+          <IconButton
+            onClick={handleCloseLogIn}
+            sx={{
+              position: "absolute",
+              right: 8,
+              top: 8,
+            }}
+          >
+            <CloseIcon />
+          </IconButton>
+        </DialogTitle>
+        <DialogContent>
+          <div className="Auth">
+            {registered ? (
+              havePass ? (
+                <LogIn
+                  handleRegistered={handleRegistered}
+                  handleHavePass={handleHavePass}
+                  handleCloseLogIn={handleCloseLogIn}
+                />
+              ) : resetPass ? (
+                <ResetPassword handleHavePass={handleHavePass} />
+              ) : (
+                <ForgetPassword handleResetPass={handleResetPass} />
+              )
+            ) : (
+              <Register handleRegistered={handleRegistered} />
+            )}
+          </div>
+        </DialogContent>
+      </Dialog>
     </div>
   );
 }
